WL-0MKYGWFDI19HQJC9: add unit tests for src/tui/id-utils.ts
